♻️ Refactor: api 변동 사항 적용
