+++ src/features/on-boarding/apis/post/create-workspace.ts
@@ -14,9 +14,17 @@ export const createWorkspace = async ({
   image,
 }: CreateWorkspaceRequestDto): Promise<BaseResponseDto<CreateWorkspaceResponseDto>> => {
   const formData = new FormData();
-  formData.append('name', name);
+
+  formData.append('request', JSON.stringify({ title: name }));
+
   if (image) {
     formData.append('image', image);
+  } else {
+    formData.append('image', '');
+  }
+
+  for (const [key, value] of formData.entries()) {
+    console.log(`[FormData] ${key}:`, value);
   }
 
   const response = await defaultApi<BaseResponseDto<CreateWorkspaceResponseDto>>(
