+++ src/features/on-boarding/constants/end-point.constants.ts
@@ -1,4 +1,5 @@
 export const API_ENDPOINTS = {
   WORKSPACES: '/workspaces',
   INVITE_MEMBERS: '/workspaces/invite',
+  WORKSPACE_DETAIL: (workspaceId: number) => `/workspace/${workspaceId}`,
 } as const;
